# improve audit_formula_name performance