Before:
```
$ time brew audit --strict -D
...
audit_formula_name: 112.98188299999998
brew audit --strict -D  146.07s user 3.34s system 98% cpu 2:31.01 total
```

After:
```
$ time brew audit --strict -D
...
audit_formula_name: 8.62922100000002
brew audit --strict -D  37.08s user 1.34s system 99% cpu 38.484 total
```